mlops/APIs/doc-ocr/requirements.txt
beautifulsoup4==4.14.2
EbookLib==0.20
fastapi==0.113.0
langdetect==1.0.9
odfpy==1.4.1
openpyxl==3.1.5
pandas==2.3.3
pdf2image==1.17.0
pillow==10.4.0
pillow_heif==0.22.0
pypdf==6.1.2
PyPDF2==3.0.1
python-docx==1.1.2
python-dotenv==1.0.1
python-multipart==0.0.9
python-pptx==1.0.2
requests==2.32.3
uvicorn==0.30.6
